% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEdges.R
\name{addEdges}
\alias{addEdges}
\title{Add edges to graph.}
\usage{
addEdges(graph, generator, type = "all", ...)
}
\arguments{
\item{graph}{[\code{grapherator}]\cr
Graph.}

\item{generator}{[\code{function(...)}]\cr
Method applied to \code{graph} in order to determine which edges to keep.
Possible values are \dQuote{onion}, \dQuote{delauney}, \dQuote{wayman} or \dQuote{grid}.}

\item{type}{[\code{character(1)}]\cr
Value \dQuote{all} applies \code{generator} to all nodes. Value \dQuote{intracluster}
instead applies the method for each cluster separately. Lastly, value \dQuote{intercluster}
applies \code{generator} to the cluster centers exclusively.
Default is \dQuote{all}.}

\item{...}{[any]\cr
Further arguments passed down to edge generator \code{generator}.}
}
\value{
[\code{grapherator}] Graph.
}
\description{
This method allows to add edges to a \code{grapherator} graph.
The method can be applied multiple times with different parameterizations. E.g.,
add edges in clusters first and add edges between clusters in a second step.
}
\seealso{
Other graph generators: \code{\link{addNodes}},
  \code{\link{addWeights}}, \code{\link{graph}}
}
