% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators-node.R
\name{coordGenerators}
\alias{coordGenerators}
\alias{addNodesLHS}
\alias{addNodesUniform}
\alias{addNodesTriangular}
\alias{addNodesGrid}
\alias{addNodesNormal}
\title{Coordinate generators.}
\usage{
addNodesLHS(n, lower = 0, upper = 1, method = NULL)

addNodesUniform(n, lower, upper)

addNodesTriangular(n, lower, upper)

addNodesGrid(n, lower, upper)

addNodesNormal(n, lower, upper, x.mean, x.sd, y.mean, y.sd)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of points to generate.}

\item{lower}{[\code{numeric(2)}]\cr
Minimal values for the first and second coordinates respectively.
Default is 0.}

\item{upper}{[\code{numeric(2)}]\cr
Maximal values for the first and second coordinates respectively.
Default is 1.}

\item{method}{[\code{function}]\cr
Function from package \pkg{lhs}.
Default is \code{\link[lhs]{maximinLHS}}.}

\item{x.mean}{[\code{numeric}]\cr
Mean value of normal distribution for x-value generation.
Only relevant for \code{\link{addNodesNormal}}.}

\item{x.sd}{[\code{numeric}]\cr
Standard deviation of normal distribution for x-value generation.
Only relevant for \code{\link{addNodesNormal}}.}

\item{y.mean}{[\code{numeric}]\cr
Mean value of normal distribution for y-value generation.
Only relevant for \code{\link{addNodesNormal}}.}

\item{y.sd}{[\code{numeric}]\cr
Standard deviation of normal distribution for y-value generation.
Only relevant for \code{\link{addNodesNormal}}.}
}
\value{
[\code{matrix(n, 2)}] Matrix of node coordinates.
}
\description{
Functions for the placement of node coordinates in the
euclidean plane. Function \code{addNodesLHS} generates a space-filling
latin hypercube sample, \code{addNodesUniform} samples points from a
bivariate uniform distribution, \code{addNodesGrid} generates a regular
grid of points, \code{addNodesTriangular} generates a regular triangular
grid and \code{addNodesNormal} generates nodes on basis of a normal
distribution.
}
