% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapherator-package.R
\docType{package}
\name{grapherator-package}
\alias{grapherator-package}
\title{grapherator: A Toolbox for the Multi-Criteria Minimum Spanning Tree Problem.}
\description{
Due to lack of real world graphs, e.g., the optimization community often
relies on artificial graph problems. The \pkg{grapherator} package implements
a multi-step approach for the generation of weighted graphs.
...
}
\section{Generation philosophy}{

...
}

\references{
[1] Bossek, J., and Grimme, C. 2017. A Pareto-Beneficial Sub-Tree Mutation
for the Multi-Criteria Minimum Spanning Tree Problem. In Proceedings of the
2017 IEEE Symposium Series on Computational Intelligence. (accepted)
}
