% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{mcGP}
\alias{mcGP}
\title{Generate a bare multi-objective graph.}
\usage{
mcGP(lower, upper)
}
\arguments{
\item{lower}{[\code{integer(1)}]\cr
Lower bounds for coordinates.}

\item{upper}{[\code{integer(1)}]\cr
Upper bounds for coordinates.}
}
\value{
[\code{mcGP}] Multi-objective graph problem.
}
\description{
This function generates a bare multi-objective graph. The generated
object does not contain nodes, edges or edge weights. It serves as a starting
point for a three step-approach of multi-objective graph problem construction:
1) Add nodes respectively coordinates via \code{\link{addCoordinates}}, add edges
via \code{\link{addEdges}} and finally add edge weights with the function
\code{\link{addWeights}}.
}
\seealso{
Other graph generators: \code{\link{addCoordinates}},
  \code{\link{addEdges}}, \code{\link{addWeights}}
}
