% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEdges.R
\name{addEdges}
\alias{addEdges}
\title{Add edges to graph.}
\usage{
addEdges(graph, generator, type = "all", ...)
}
\arguments{
\item{graph}{[\code{\link{grapherator}}]\cr
Graph.}

\item{generator}{[\code{function(graph, ...)}]\cr
Method applied to \code{graph} in order to determine which edges to add.}

\item{type}{[\code{character(1)}]\cr
Value \dQuote{all} applies \code{generator} to all nodes. Value \dQuote{intracluster}
instead applies the method for each cluster separately. Lastly, value \dQuote{intercluster}
applies \code{generator} to the cluster centers exclusively.
Default is \dQuote{all}.}

\item{...}{[any]\cr
Further arguments passed down to edge generator \code{generator}.}
}
\value{
[\code{\link{grapherator}}] Graph.
}
\description{
This method allows to add edges to a \code{grapherator} graph.
The method can be applied multiple times with different parameterizations. E.g.,
add edges in clusters first and add edges between clusters in a second step.
}
\references{
Erdos, P., and A. Renyi. 1959. "On random graphs, I." Publicationes Mathematicae
(Debrecen) 6: 290-97.

Waxman, B. M. 1988. "Routing of Multipoint Connections."" IEEE Journal on Selected
Areas in Communications 6 (9): 1617-22. doi:10.1109/49.12889.

Knowles, J. D., and D. W. Corne. 2001. "Benchmark Problem Generators and Results
for the Multiobjective Degree-Constrained Minimum Spanning Tree Problem."
In Proceedings of the 3rd Annual Conference on Genetic and Evolutionary
Computation, 424-31. GECCO'01. San Francisco, CA, USA: Morgan Kaufmann Publishers Inc.
}
\seealso{
Other graph generators: \code{\link{addNodes}},
  \code{\link{addWeights}}, \code{\link{graph}}
}
