% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{addEdges}
\alias{addEdges}
\title{Define edges in multi-objective graph.}
\usage{
addEdges(graph, method, type = "all", ...)
}
\arguments{
\item{graph}{[\code{mcGP}]\cr
Multi-objective graph problem.}

\item{method}{[\code{function(...)}]\cr
Method applied to \code{graph} in order to determine which edges to keep.
Possible values are \dQuote{onion}, \dQuote{delauney}, \dQuote{wayman} or \dQuote{grid}.}

\item{type}{[\code{character(1)}]\cr
Value \dQuote{all} applies \code{method} to all nodes. Value \dQuote{intreacluster}
instead applies the method for each cluster separately. Lastly, value \dQuote{intercluster}
applies \code{method} to the cluster centers.}

\item{...}{[any]\cr
Passed down to edge constructor.}
}
\value{
[\code{mcGP}] Multi-objective graph problem.
}
\description{
By default \code{\link{addWeights}} generates n(n-1)/2 weights, i.e.,
the graph is assumed to be complete. This method allows to defne an adjacency
matrix to make the graph more sparse. The method can be applied multiple times
with different parameterizations.
}
\seealso{
Other graph generators: \code{\link{addCoordinates}},
  \code{\link{addWeights}}, \code{\link{mcGP}}
}
