% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeights.R
\name{addWeights}
\alias{addWeights}
\title{Add weights to graph.}
\usage{
addWeights(graph, generator = NULL, weights = NULL, symmetric = TRUE,
  to.int = FALSE, ...)
}
\arguments{
\item{graph}{[\code{grapherator}]\cr
Graph.}

\item{generator}{[\code{function(graph, ...)}]\cr
Function used to generate weights. The functions needs to expect the graph
as the first argument \code{graph}. Additional control argument are
possible.}

\item{weights}{[\code{matrix}]\cr
Square matrix of weights.
If some weights are already assigned, pay attention to the correct dimensions.
If this is passed all other arguments are ignored.
Default is \code{NULL}.}

\item{symmetric}{[\code{logical(1)}]\cr
Should the weights be symmetric, i.e., \eqn{w(i, j) = w(j, i)} for each pair \eqn{i, j} of nodes?
Default is \code{TRUE}.}

\item{to.int}{[\code{logical(1)}]\cr
Should weights be rounded to integer?
Default is \code{FALSE}.}

\item{...}{[any]\cr
Additional arguments passed down to \code{generator}.}
}
\value{
[\code{grapherator}] Graph.
}
\description{
\code{addWeights} allows to add edge weights to a graph. This is
the last step of the graph generation process. Note that adding edges is not
possible once \code{addWeights} was called.
The weights can be generated on basis of the node coordinates (in this
case \code{\link[stats]{dist}} is applied with the cooresponding \code{method}).
Alternatively, all kinds of random weights can be generated.
}
\seealso{
Other graph generators: \code{\link{addEdges}},
  \code{\link{addNodes}}, \code{\link{graph}}
}
