% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Generate a bare graph.}
\usage{
graph(lower, upper)
}
\arguments{
\item{lower}{[\code{integer(1)}]\cr
Lower bounds for node coordinates in the Euclidean plane.}

\item{upper}{[\code{integer(1)}]\cr
Upper bounds for node coordinates in the Euclidean plane.}
}
\value{
[\code{grapherator}] Graph.
}
\description{
This function generates a bare graph. The generated
object does not contain nodes, edges or edge weights. It serves as a starting
point for a three step approach of graph problem construction:
1) Add nodes respectively coordinates via \code{\link{addNodes}}, add edges
via \code{\link{addEdges}} and finally add edge weights with the function
\code{\link{addWeights}}.
}
\seealso{
Other graph generators: \code{\link{addEdges}},
  \code{\link{addNodes}}, \code{\link{addWeights}}
}
