% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.grapherator}
\alias{plot.grapherator}
\title{Visualize bi-objective graph.}
\usage{
\method{plot}{grapherator}(x, y = NULL, show.cluster.centers = TRUE,
  highlight.clusters = FALSE, show.edges = TRUE, ...)
}
\arguments{
\item{x}{[\code{grapherator}]\cr
Multi-objective graph problem.}

\item{y}{Not used at the moment.}

\item{show.cluster.centers}{[\code{logical(1)}]\cr
Display cluster centers?
Default is \code{TRUE}. This option is ignored silently if the instance is not clustered.}

\item{highlight.clusters}{[\code{logical(1)}]\cr
Shall nodes be coloured by cluster membership?
Default is \code{FALSE}.}

\item{show.edges}{[\code{logical(1)}]\cr
Display edges? Keep in mind, that the number of edges is \eqn{O(n^2)}
where \eqn{n} is the number of nodes.
Default is \code{TRUE}.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{list}] A list of \code{\link[ggplot2]{ggplot}} objects with components
\code{pl.weights} (scatterplot of edge weights) and eventually \code{pl.coords} (scatterplot of
nodes). The latter is \code{NULL}, if \code{graph} has no associated coordinates.
}
\description{
Only applicable for bi-objective problems of class \code{grapherator}.
\code{plot.grapherator} generates a scatterplot of edge weights. If the nodes do
have coordinates, additionally a scatterplot of the nodes in the euclidean
plane is generated.
}
